add docs age on TM-BOM and rename test file
