+++ td.vue/tests/unit/service/migration/tmBom/diagrams/nodes.spec.js
@@ -1,5 +1,5 @@
 import nodes from '@/service/migration/tmBom/diagrams/nodes';
-import tmBomModel from '../husky-ai-threat-model';
+import tmBomModel from '../test-model';
 
 describe('service/migration/tmBom/diagrams/nodes.js', () => {
