+++ td.vue/tests/unit/service/migration/tmBom/summary.spec.js
@@ -1,5 +1,5 @@
 import summary from '@/service/migration/tmBom/summary';
-import tmBomModel from './husky-ai-threat-model';
+import tmBomModel from './test-model';
 import tdModel from './v2-threat-model';
 
 describe('service/migration/tmBom/summary.js', () => {
